Atualiza tipos de dados e simplifica lógica de sincronização
As alterações incluem a atualização de tipos de dados de `int` para `long` em entidades como distritos, mesorregiões, microrregiões, municípios, regiões e estados, garantindo maior intervalo de valores para IDs.

Os manipuladores de comandos foram adaptados para utilizar novos métodos de repositório que aceitam IDs do tipo `long`, simplificando a lógica de sincronização com listas de novos itens.

As interfaces de repositório foram atualizadas para refletir essas mudanças e incluir métodos para adição em massa de entidades.

As classes de DTO também foram alteradas para usar `long` para IDs, mantendo a consistência com as entidades do domínio.

O `IbgeService` foi modificado para usar `Newtonsoft.Json` para a desserialização, oferecendo mais flexibilidade.

Adicionadas migrações para atualizar o banco de dados, refletindo as novas definições de classe.

+++ Core/Application/Interfaces/Repositories/IDistrictsRepository.cs
@@ -5,6 +5,6 @@ namespace Application.Interfaces.Repositories;
 public interface IDistrictsRepository : IRepository<Districts>
 {
     Task<List<Districts>> GetAllAsync();
-    Task<Districts> GetByIdAsync(int id);
-    Task<List<Districts>> GetByMunicipalityIdAsync(int municipalityId);
+    Task<Districts> GetByIdAsync(long id);
+    Task<List<Districts>> GetByMunicipalityIdAsync(long municipalityId);
 }
