+++ Core/Application/Interfaces/Repositories/ISubDistrictsRepository.cs
@@ -5,6 +5,6 @@ namespace Application.Interfaces.Repositories;
 public interface ISubDistrictsRepository : IRepository<SubDistricts>
 {
     Task<List<SubDistricts>> GetAllAsync();
-    Task<SubDistricts> GetByIdAsync(int id);
-    Task<List<SubDistricts>> GetByDistrictIdAsync(int districtId);
+    Task<SubDistricts> GetByIdAsync(long id);
+    Task<List<SubDistricts>> GetByDistrictIdAsync(long districtId);
 }
