+++ Infraestructure/Elasticsearch/DTOs/MunicipalityDto.cs
@@ -2,7 +2,7 @@
 
 public class MunicipalityDto
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; } = null!;
-    public int MicroRegionId { get; set; }
+    public long MicroRegionId { get; set; }
 }
